+++ apps/ethereum_jsonrpc/lib/ethereum_jsonrpc/parity.ex
@@ -49,6 +49,24 @@ defmodule EthereumJSONRPC.Parity do
     end
   end
 
+  @impl EthereumJSONRPC.Variant
+  def fetch_first_trace(transactions_params, json_rpc_named_arguments) when is_list(transactions_params) do
+    id_to_params = id_to_params(transactions_params)
+
+    with {:ok, responses} <-
+           id_to_params
+           |> trace_replay_transaction_requests()
+           |> json_rpc(json_rpc_named_arguments) do
+      {:ok, [first_trace]} = trace_replay_transaction_responses_to_first_trace_params(responses, id_to_params)
+
+      %{block_hash: block_hash} =
+        transactions_params
+        |> Enum.at(0)
+
+      {:ok, [%{first_trace: first_trace, block_hash: block_hash, json_rpc_named_arguments: json_rpc_named_arguments}]}
+    end
+  end
+
   @doc """
   Fetches the pending transactions from the Parity node.
 
@@ -166,5 +184,110 @@ defmodule EthereumJSONRPC.Parity do
 
   defp trace_replay_block_transactions_request(%{id: id, block_number: block_number}) do
     request(%{id: id, method: "trace_replayBlockTransactions", params: [integer_to_quantity(block_number), ["trace"]]})
+  end
+
+  def trace_replay_transaction_responses_to_first_trace_params(responses, id_to_params)
+      when is_list(responses) and is_map(id_to_params) do
+    with {:ok, traces} <- trace_replay_transaction_responses_to_first_trace(responses, id_to_params) do
+      params =
+        traces
+        |> Traces.to_elixir()
+        |> Traces.elixir_to_params()
+
+      {:ok, params}
+    end
+  end
+
+  defp trace_replay_transaction_responses_to_first_trace(responses, id_to_params)
+       when is_list(responses) and is_map(id_to_params) do
+    responses
+    |> Enum.map(&trace_replay_transaction_response_to_first_trace(&1, id_to_params))
+    |> Enum.reduce(
+      {:ok, []},
+      fn
+        {:ok, traces}, {:ok, acc_traces_list} ->
+          {:ok, [traces | acc_traces_list]}
+
+        {:ok, _}, {:error, _} = acc_error ->
+          acc_error
+
+        {:error, reason}, {:ok, _} ->
+          {:error, [reason]}
+
+        {:error, reason}, {:error, acc_reason} ->
+          {:error, [reason | acc_reason]}
+      end
+    )
+    |> case do
+      {:ok, traces_list} ->
+        traces =
+          traces_list
+          |> Enum.reverse()
+          |> List.flatten()
+
+        {:ok, traces}
+
+      {:error, reverse_reasons} ->
+        reasons = Enum.reverse(reverse_reasons)
+        {:error, reasons}
+    end
+  end
+
+  defp trace_replay_transaction_response_to_first_trace(%{id: id, result: %{"trace" => traces}}, id_to_params)
+       when is_list(traces) and is_map(id_to_params) do
+    %{
+      block_hash: block_hash,
+      block_number: block_number,
+      hash_data: transaction_hash,
+      transaction_index: transaction_index
+    } = Map.fetch!(id_to_params, id)
+
+    first_trace =
+      traces
+      |> Stream.with_index()
+      |> Enum.map(fn {trace, index} ->
+        Map.merge(trace, %{
+          "blockHash" => block_hash,
+          "blockNumber" => block_number,
+          "index" => index,
+          "transactionIndex" => transaction_index,
+          "transactionHash" => transaction_hash
+        })
+      end)
+      |> Enum.filter(fn trace ->
+        Map.get(trace, "index") == 0
+      end)
+
+    {:ok, first_trace}
+  end
+
+  defp trace_replay_transaction_response_to_first_trace(%{id: id, error: error}, id_to_params)
+       when is_map(id_to_params) do
+    %{
+      block_hash: block_hash,
+      block_number: block_number,
+      hash_data: transaction_hash,
+      transaction_index: transaction_index
+    } = Map.fetch!(id_to_params, id)
+
+    annotated_error =
+      Map.put(error, :data, %{
+        "blockHash" => block_hash,
+        "blockNumber" => block_number,
+        "transactionIndex" => transaction_index,
+        "transactionHash" => transaction_hash
+      })
+
+    {:error, annotated_error}
+  end
+
+  defp trace_replay_transaction_requests(id_to_params) when is_map(id_to_params) do
+    Enum.map(id_to_params, fn {id, %{hash_data: hash_data}} ->
+      trace_replay_transaction_request(%{id: id, hash_data: hash_data})
+    end)
+  end
+
+  defp trace_replay_transaction_request(%{id: id, hash_data: hash_data}) do
+    request(%{id: id, method: "trace_replayTransaction", params: [hash_data, ["trace"]]})
   end
 end